remove src/validations/yups/schema.tsx
